server/requirements.txt
amqp==5.1.1
anyio==3.6.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
dnspython==2.2.1
fastapi==0.87.0
h11==0.14.0
httpcore==0.16.2
httptools==0.5.0
httpx==0.23.1
idna==3.4
kombu==5.2.4
motor==3.1.1
prompt-toolkit==3.0.33
pydantic==1.10.2
pymongo==4.3.3
python-dotenv==0.21.0
pytz==2022.6
PyYAML==6.0
requests==2.28.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
starlette==0.21.0
typing_extensions==4.4.0
urllib3==1.26.13
uvicorn==0.20.0
uvloop==0.17.0
vine==5.0.0
watchfiles==0.18.1
wcwidth==0.2.5
websockets==10.4
xmltodict==0.13.0
